♻️Refactor(GATHERING): 모임글 상세보기의 반환클래스 수정 및 교체
- 기존의 상세보기의 반환 클래스는 GatheringResponse 였고, 동일한 이름의 모임목록 조회시의 반환클래스가 있는것을 발견(dto 가 아닌 type 패키지에 잘못 저장했었음)
- GatheringResponse의 이름을 모임목록에 사용할 dto로 변경
- 상세보기에 사용할 GatheringDetailContentResponse 새로 생성
   - 모임글 상세보기의 반환클래스: GatheringResponse -> GatheringDetailContentResponse 로 교체
- type 에 잘못 저장했던 GatheringResponse 삭제

Related to: #24
